rename /verifyReceipt as /verify
